+++ domain-20180208/alibabacloud_domain20180208/models.py
@@ -112,6 +112,288 @@ class AcceptDemandResponse(TeaModel):
         return self
 
 
+class BatchIntrudeDomainsRequest(TeaModel):
+    def __init__(
+        self,
+        domain_names: List[str] = None,
+    ):
+        self.domain_names = domain_names
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain_names is not None:
+            result['DomainNames'] = self.domain_names
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DomainNames') is not None:
+            self.domain_names = m.get('DomainNames')
+        return self
+
+
+class BatchIntrudeDomainsShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        domain_names_shrink: str = None,
+    ):
+        self.domain_names_shrink = domain_names_shrink
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain_names_shrink is not None:
+            result['DomainNames'] = self.domain_names_shrink
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DomainNames') is not None:
+            self.domain_names_shrink = m.get('DomainNames')
+        return self
+
+
+class BatchIntrudeDomainsResponseBodyDataFailureList(TeaModel):
+    def __init__(
+        self,
+        domain_name: str = None,
+        error_code: str = None,
+        error_msg: str = None,
+        success: bool = None,
+    ):
+        self.domain_name = domain_name
+        self.error_code = error_code
+        self.error_msg = error_msg
+        self.success = success
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain_name is not None:
+            result['DomainName'] = self.domain_name
+        if self.error_code is not None:
+            result['ErrorCode'] = self.error_code
+        if self.error_msg is not None:
+            result['ErrorMsg'] = self.error_msg
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DomainName') is not None:
+            self.domain_name = m.get('DomainName')
+        if m.get('ErrorCode') is not None:
+            self.error_code = m.get('ErrorCode')
+        if m.get('ErrorMsg') is not None:
+            self.error_msg = m.get('ErrorMsg')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class BatchIntrudeDomainsResponseBodyDataSuccessList(TeaModel):
+    def __init__(
+        self,
+        domain_name: str = None,
+        error_code: str = None,
+        error_msg: str = None,
+        success: bool = None,
+    ):
+        self.domain_name = domain_name
+        self.error_code = error_code
+        self.error_msg = error_msg
+        self.success = success
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain_name is not None:
+            result['DomainName'] = self.domain_name
+        if self.error_code is not None:
+            result['ErrorCode'] = self.error_code
+        if self.error_msg is not None:
+            result['ErrorMsg'] = self.error_msg
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DomainName') is not None:
+            self.domain_name = m.get('DomainName')
+        if m.get('ErrorCode') is not None:
+            self.error_code = m.get('ErrorCode')
+        if m.get('ErrorMsg') is not None:
+            self.error_msg = m.get('ErrorMsg')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class BatchIntrudeDomainsResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        failure_count: int = None,
+        failure_list: List[BatchIntrudeDomainsResponseBodyDataFailureList] = None,
+        success_count: int = None,
+        success_list: List[BatchIntrudeDomainsResponseBodyDataSuccessList] = None,
+    ):
+        self.failure_count = failure_count
+        self.failure_list = failure_list
+        self.success_count = success_count
+        self.success_list = success_list
+
+    def validate(self):
+        if self.failure_list:
+            for k in self.failure_list:
+                if k:
+                    k.validate()
+        if self.success_list:
+            for k in self.success_list:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.failure_count is not None:
+            result['FailureCount'] = self.failure_count
+        result['FailureList'] = []
+        if self.failure_list is not None:
+            for k in self.failure_list:
+                result['FailureList'].append(k.to_map() if k else None)
+        if self.success_count is not None:
+            result['SuccessCount'] = self.success_count
+        result['SuccessList'] = []
+        if self.success_list is not None:
+            for k in self.success_list:
+                result['SuccessList'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('FailureCount') is not None:
+            self.failure_count = m.get('FailureCount')
+        self.failure_list = []
+        if m.get('FailureList') is not None:
+            for k in m.get('FailureList'):
+                temp_model = BatchIntrudeDomainsResponseBodyDataFailureList()
+                self.failure_list.append(temp_model.from_map(k))
+        if m.get('SuccessCount') is not None:
+            self.success_count = m.get('SuccessCount')
+        self.success_list = []
+        if m.get('SuccessList') is not None:
+            for k in m.get('SuccessList'):
+                temp_model = BatchIntrudeDomainsResponseBodyDataSuccessList()
+                self.success_list.append(temp_model.from_map(k))
+        return self
+
+
+class BatchIntrudeDomainsResponseBody(TeaModel):
+    def __init__(
+        self,
+        data: BatchIntrudeDomainsResponseBodyData = None,
+        request_id: str = None,
+    ):
+        self.data = data
+        self.request_id = request_id
+
+    def validate(self):
+        if self.data:
+            self.data.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.data is not None:
+            result['Data'] = self.data.to_map()
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Data') is not None:
+            temp_model = BatchIntrudeDomainsResponseBodyData()
+            self.data = temp_model.from_map(m['Data'])
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        return self
+
+
+class BatchIntrudeDomainsResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: BatchIntrudeDomainsResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = BatchIntrudeDomainsResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class BidDomainRequest(TeaModel):
     def __init__(
         self,
@@ -3419,6 +3701,167 @@ class QueryExchangeRateResponse(TeaModel):
         return self
 
 
+class QueryExportAuctionDetailRequest(TeaModel):
+    def __init__(
+        self,
+        auction_id: str = None,
+    ):
+        self.auction_id = auction_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.auction_id is not None:
+            result['AuctionId'] = self.auction_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AuctionId') is not None:
+            self.auction_id = m.get('AuctionId')
+        return self
+
+
+class QueryExportAuctionDetailResponseBody(TeaModel):
+    def __init__(
+        self,
+        auction_end_time: str = None,
+        auction_id: str = None,
+        auction_status: str = None,
+        book_end_time: str = None,
+        buyer_status: str = None,
+        current_price: float = None,
+        increase_price: float = None,
+        my_price: float = None,
+        my_proxy_price: float = None,
+        others_max_proxy_price: float = None,
+        proxy_price: float = None,
+        request_id: str = None,
+    ):
+        self.auction_end_time = auction_end_time
+        self.auction_id = auction_id
+        self.auction_status = auction_status
+        self.book_end_time = book_end_time
+        self.buyer_status = buyer_status
+        self.current_price = current_price
+        self.increase_price = increase_price
+        self.my_price = my_price
+        self.my_proxy_price = my_proxy_price
+        self.others_max_proxy_price = others_max_proxy_price
+        self.proxy_price = proxy_price
+        self.request_id = request_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.auction_end_time is not None:
+            result['AuctionEndTime'] = self.auction_end_time
+        if self.auction_id is not None:
+            result['AuctionId'] = self.auction_id
+        if self.auction_status is not None:
+            result['AuctionStatus'] = self.auction_status
+        if self.book_end_time is not None:
+            result['BookEndTime'] = self.book_end_time
+        if self.buyer_status is not None:
+            result['BuyerStatus'] = self.buyer_status
+        if self.current_price is not None:
+            result['CurrentPrice'] = self.current_price
+        if self.increase_price is not None:
+            result['IncreasePrice'] = self.increase_price
+        if self.my_price is not None:
+            result['MyPrice'] = self.my_price
+        if self.my_proxy_price is not None:
+            result['MyProxyPrice'] = self.my_proxy_price
+        if self.others_max_proxy_price is not None:
+            result['OthersMaxProxyPrice'] = self.others_max_proxy_price
+        if self.proxy_price is not None:
+            result['ProxyPrice'] = self.proxy_price
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AuctionEndTime') is not None:
+            self.auction_end_time = m.get('AuctionEndTime')
+        if m.get('AuctionId') is not None:
+            self.auction_id = m.get('AuctionId')
+        if m.get('AuctionStatus') is not None:
+            self.auction_status = m.get('AuctionStatus')
+        if m.get('BookEndTime') is not None:
+            self.book_end_time = m.get('BookEndTime')
+        if m.get('BuyerStatus') is not None:
+            self.buyer_status = m.get('BuyerStatus')
+        if m.get('CurrentPrice') is not None:
+            self.current_price = m.get('CurrentPrice')
+        if m.get('IncreasePrice') is not None:
+            self.increase_price = m.get('IncreasePrice')
+        if m.get('MyPrice') is not None:
+            self.my_price = m.get('MyPrice')
+        if m.get('MyProxyPrice') is not None:
+            self.my_proxy_price = m.get('MyProxyPrice')
+        if m.get('OthersMaxProxyPrice') is not None:
+            self.others_max_proxy_price = m.get('OthersMaxProxyPrice')
+        if m.get('ProxyPrice') is not None:
+            self.proxy_price = m.get('ProxyPrice')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        return self
+
+
+class QueryExportAuctionDetailResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: QueryExportAuctionDetailResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = QueryExportAuctionDetailResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class QueryExportDomainExpireSnatchsRequest(TeaModel):
     def __init__(
         self,
@@ -3498,11 +3941,13 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
         product_id: str = None,
         publish_time: str = None,
         reg_date: str = None,
+        renew_price: float = None,
         reserved: bool = None,
         rmb_price: float = None,
         s_360weight: int = None,
         seo_attributes: str = None,
         short_name: str = None,
+        snatch_no: str = None,
         snatch_type_desc: str = None,
         sougou_anti_link: int = None,
         sougou_index: int = None,
@@ -3538,11 +3983,13 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
         self.product_id = product_id
         self.publish_time = publish_time
         self.reg_date = reg_date
+        self.renew_price = renew_price
         self.reserved = reserved
         self.rmb_price = rmb_price
         self.s_360weight = s_360weight
         self.seo_attributes = seo_attributes
         self.short_name = short_name
+        self.snatch_no = snatch_no
         self.snatch_type_desc = snatch_type_desc
         self.sougou_anti_link = sougou_anti_link
         self.sougou_index = sougou_index
@@ -3615,6 +4062,8 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
             result['PublishTime'] = self.publish_time
         if self.reg_date is not None:
             result['RegDate'] = self.reg_date
+        if self.renew_price is not None:
+            result['RenewPrice'] = self.renew_price
         if self.reserved is not None:
             result['Reserved'] = self.reserved
         if self.rmb_price is not None:
@@ -3625,6 +4074,8 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
             result['SeoAttributes'] = self.seo_attributes
         if self.short_name is not None:
             result['ShortName'] = self.short_name
+        if self.snatch_no is not None:
+            result['SnatchNo'] = self.snatch_no
         if self.snatch_type_desc is not None:
             result['SnatchTypeDesc'] = self.snatch_type_desc
         if self.sougou_anti_link is not None:
@@ -3697,6 +4148,8 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
             self.publish_time = m.get('PublishTime')
         if m.get('RegDate') is not None:
             self.reg_date = m.get('RegDate')
+        if m.get('RenewPrice') is not None:
+            self.renew_price = m.get('RenewPrice')
         if m.get('Reserved') is not None:
             self.reserved = m.get('Reserved')
         if m.get('RmbPrice') is not None:
@@ -3707,6 +4160,8 @@ class QueryExportDomainExpireSnatchsResponseBodyData(TeaModel):
             self.seo_attributes = m.get('SeoAttributes')
         if m.get('ShortName') is not None:
             self.short_name = m.get('ShortName')
+        if m.get('SnatchNo') is not None:
+            self.snatch_no = m.get('SnatchNo')
         if m.get('SnatchTypeDesc') is not None:
             self.snatch_type_desc = m.get('SnatchTypeDesc')
         if m.get('SougouAntiLink') is not None: